Back-end/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.6
django-admin-rangefilter==0.10.0
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.10.0
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.59.7
packaging==24.2
parso==0.8.4
pathlib==1.0.1
Pillow==9.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2023.3
pywin32==308
pyzmq==26.2.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
spotipy==2.25.0
sqlparse==0.5.1
stack-data==0.6.3
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Wikipedia-API==0.7.3
zipp==3.21.0
